Do not mount /etc/passwd and /etc/group files (#162)
The goal is of this volume mount was to translate usernames from ids. However this is not really necessary, and exposes the full content of the passwd and group file from the host to the container.
Moreover, on systems like MacOS where docker runs into an internal Linux VM, the content of the file /etc/passwd is completely irrelevant.